Big Changes:
Added the option to not load in volume data and instead read them out of the unpacked data containers on the fly. This is generally very slow. Therefore added option to temporarily load in a whole ForceMap for batch processing.

Needed to replace all instances of volume data ForceMap properties in the code with a new data read out method '[OutForce,OutHeight] = get_force_curve_data'