Added possibility to just invert plot axes of some special case Equations in numerically_solve_bivariate_equation_into_lookup_table to
Major speedupfor equations were the independent is isolated on one side. Can not control the independents range this way. Instead, the dependents range must be given and all else is determined from that.

Also, if the Eq is in its solved form already, the function now just inserts values from RangeIndependent instead of calling vpasolve