Added check_and_assert_channel_image_sizes
This will ensure that the new canges from previous the commit will apply to Experiments that have been created on the older version.

Caution: This does not fix every possible problem that derived from the NumPixel bug. It mostly makes all the original channels display properly in the show_image viewer. If in doubt, start the Experiment from scratch.